+++ tests/test_hierarchical_transformer.py
@@ -28,6 +28,7 @@ def test_hierarchical_transformer():
             stride=4,
             padding=2,
             seq_len=image_size * image_size // 16,
+            feedforward="MLP",
         ),
         BasicLayerConfig(
             embedding=128,
@@ -36,6 +37,7 @@ def test_hierarchical_transformer():
             stride=2,
             padding=1,
             seq_len=image_size * image_size // 64,
+            feedforward="MLP",
         ),
         BasicLayerConfig(
             embedding=320,
@@ -44,6 +46,7 @@ def test_hierarchical_transformer():
             stride=2,
             padding=1,
             seq_len=image_size * image_size // 256,
+            feedforward="MLP",
         ),
     ]